Updated directory structure
