+++ README.md
@@ -1,2 +1,2 @@
 # ImmutableSOC
-A library that contains all SOC state information and can control an SOC game. The state objects are all immutable to control state transitions. 
+A library that contains all SOC soc.state information and can control an SOC game. The soc.state objects are all immutable to control soc.state transitions.
