--- src/main/scala/core/Roll.scala
@@ -1,6 +1,0 @@
-package core
-
-case class Roll(number: Int) {
-  val dots: Int = 6 - Math.abs(7 - number)
-  val prob: Double = dots.toDouble / 36.0
-}
