--- src/main/scala/dice/Dice.scala
@@ -1,7 +1,0 @@
-package dice
-
-trait Dice {
-
-  def getRoll: (Int, Int)
-
-}
