+++ src/main/scala/soc/dice/Dice.scala
@@ -1,0 +1,7 @@
+package soc.dice
+
+trait Dice {
+
+  def getRoll: (Int, Int)
+
+}
