+++ src/main/scala/soc/state/TurnState.scala
@@ -1,0 +1,6 @@
+package soc.state
+
+case class TurnState(
+  canPlayDevCard: Boolean = true,
+  canRollDice: Boolean = true
+)
