--- src/main/scala/state/TurnState.scala
@@ -1,6 +1,0 @@
-package state
-
-case class TurnState(
-  canPlayDevCard: Boolean = true,
-  canRollDice: Boolean = true
-)
